filter update 4

+++ backend/KAZABUILD.Application/DTOs/Components/Components/MemoryComponent/GetMemoryComponentDto.cs
@@ -37,7 +37,7 @@ namespace KAZABUILD.Application.DTOs.Components.Components.MemoryComponent
         public int? ModuleQuantityEnd { get; set; }
 
         [Range(64, 524288, ErrorMessage = "Module Capacity must be between 64 MB and 512 GB")]
-        public decimal? ModuleCapacityStart { get; set; }
+        public int? ModuleCapacityStart { get; set; }
 
         [Range(64, 524288, ErrorMessage = "Module Capacity must be between 64 MB and 512 GB")]
         public int? ModuleCapacityEnd { get; set; }
